server/requirements.txt
fastapi==0.78.0
keras==2.8.0
nltk==3.7
numpy==1.23.0
tensorflow-cpu
gunicorn
uvicorn
